Add fallback for browsers that do not support Clipboard API
This fixes #3 by adding a fallback to `execCommand()` for browsers that do not support the Clipboard API such as Safari.

This fallback creates a `ghost` textarea with the svg code to be copied as the value. This is so we can automatically select the text and copy to clipboard using `execCommand("copy")`. This `ghost` textarea is invisible and removed immediately.